requirements.txt
allure-pytest == 2.10.0

boto3 == 1.21.46

botocore == 1.24.46

jsonpath-rw == 1.4.0

jsonpath-rw-ext == 1.2.2

pytest == 7.1.2

pytest-assume == 2.4.3

pytest-forked ==1.4.0

pytest-html == 3.1.1

pytest-check == 2.1.4

pytest-xdist == 2.5.0

python-dateutil == 2.8.2

requests==2.27.1

colorlog==6.7.0

pyyaml

jsonpath_rw_ext == 1.2.2

click == 8.1.3

pytz == 2023.3

PyJWT == 2.8.0
